Move wallsign check to post-house-sign validation
Previously the check in the house sign creation process concerning the
sign being a wall sign happened prior to actually validating that the
sign is going to be a house point related sign.

This had the effect that players with permission to create house point
signs would be notified with the wall-sign error message whenever
creating a sign that is not a wall sign itself.

To fix this, this commit moves this check after the actually validation
of the signs content to be a house points, therefore only notifying
players about an incorrect sign type if the sign itself is going to be a
house points sign.